requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.7.0
ansi2html==1.9.1
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.7.0
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
Brotli @ file:///C:/Windows/Temp/abs_63l7912z0e/croots/recipe/brotli-split_1659616056886/work
catboost @ https://pypi.org/packages/cp39/c/catboost/catboost-1.2.2-cp39-cp39-win_amd64.whl#sha256=1c0c3973542c5c46898c157a9f545ad80a36b8d8b17d92424a07376818911162
category-encoders==2.6.3
certifi==2023.7.22
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.28.3
colorama==0.4.6
colorlog==6.8.0
comm==0.1.4
contourpy==1.2.0
cramjam==2.7.0
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
Cython==3.0.6
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.19.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
email_validator==2.2.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
fastparquet==2023.10.1
fhir==0.0.4
fhir.resources==7.1.0
fhirclient==4.2.0
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
Flask==3.0.0
flatbuffers==24.3.25
focal-loss==0.0.7
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.41
google-pasta==0.2.0
graphviz @ file:///home/conda/feedstock_root/build_artifacts/python-graphviz_1658658635601/work
greenlet==3.0.2
grpcio==1.64.0
h5py==3.10.0
huggingface-hub==0.23.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.33.1
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib @ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kaleido==0.2.1
keras==3.4.1
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
lazy_loader==0.3
libclang==18.1.1
lightgbm @ file:///C:/b/abs_636j7r6zqr/croot/lightgbm_1677766516694/work
lightning==2.4.0
lightning-utilities==0.11.2
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci/markupsafe_1654508077284/work
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
ml-dtypes==0.4.0
mne==1.7.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
ndjson==0.3.1
nest-asyncio==1.5.8
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numexpr @ file:///C:/b/abs_5fucrty5dc/croot/numexpr_1696515448831/work
numpy==1.25.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
optuna==3.5.0
orjson==3.9.10
overrides==7.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandarallel==1.6.5
pandas @ file:///C:/b/abs_b9kefbuby2/croot/pandas_1677835593760/work
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peft==0.10.0
pillow==10.2.0
platformdirs==3.11.0
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1698272730927/work
plotly-resampler==0.9.1
pmdarima==2.0.4
pooch==1.8.1
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.2
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycaret==3.2.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==2.4.4
pydot @ file:///D:/bld/pydot_1695469292926/work
pydotplus @ file:///home/conda/feedstock_root/build_artifacts/pydotplus_1622588015615/work
Pygments==2.16.1
pyhealth==1.1.6
pyod==1.1.2
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.7
pytorch-lightning==2.4.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1693930252784/work
pywin32==306
pywinpty==2.0.12
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
rdkit==2023.9.5
referencing==0.30.2
regex==2023.12.25
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.12.0
safetensors==0.4.2
schemdraw==0.15
scikit-base==0.5.2
scikit-image==0.22.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
sentry-sdk==1.40.0
setproctitle==1.3.3
shap==0.44.0
shtab==1.7.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sktime==0.21.1
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
statsmodels==0.14.0
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tbats==1.1.3
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1692026804430/work
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2023.12.9
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchmetrics==1.4.0.post0
torchvision==0.19.0+cu118
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets==5.13.0
transformers==4.40.1
trl==0.8.6
tsdownsample==0.1.2
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
wandb==0.17.2
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
yellowbrick==1.5
zipp==3.17.0
